codetorunonserver

+++ FullSynthetic/DDPM-MNIST/initialmain.py
@@ -23,8 +23,8 @@ import pickle
 
 N_no_w = 60000
 N_with_w = 60000
-max_c = 2500
-Wall = [0,0.25,0.5,0.75,1,1.25,1.5, 1.75, 2]
+max_c = 500
+Wall = [0]
 generation_number = 20
 n_epoch = 40
 batch_size = 128
@@ -34,10 +34,11 @@ n_classes = 10
 n_feat = 128 # 128 ok, 256 better (but slower)
 lrate = 1e-4
 save_model = True
-save_dir_all = ['./data/t500/']
+save_dir_all = ['./data/t500/', './data/diffusion_outputs10/']
 
 
 #### Extracting mnist data statistics
+
 
 tf = transforms.Compose([transforms.ToTensor()])
 dataset = MNIST("./data", train=True, download=True, transform=tf)
@@ -49,39 +50,50 @@ cov_all = np.cov(np.transpose(real_features))
 
 ###########
 
-dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=True, num_workers=5)
+dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=True, num_workers=0)
 
 ddpm = util.DDPM(nn_model=util.ContextUnet(in_channels=1, n_feat=n_feat, n_classes=n_classes), betas=(1e-4, 0.02), n_T= n_T, device=device, drop_prob=0.1)
 ddpm.to(device)
 
-optim = torch.optim.Adam(ddpm.parameters(), lr=lrate)
-opt_loss = 1
-pbar = tqdm(range(n_epoch),leave = True)
-for ep in pbar:
-    ddpm.train()
-        
-            # linear lrate decay
-    optim.param_groups[0]['lr'] = lrate*(1-ep/n_epoch)
-        
-    loss_ema = None
-    for x, c in dataloader:
-        optim.zero_grad()
-        x = x.to(device)
-        c = c.to(device)
-        loss = ddpm(x, c, 1, 1, guide_w=0)
-        loss = loss.sum()
-        loss.backward()
-        if loss_ema is None:
-            loss_ema = loss.item()
-        else:
-            loss_ema = 0.95 * loss_ema + 0.05 * loss.item()
-            pbar.set_description(f"loss: {loss_ema:.4f}")
-            optim.step()
-                
-if save_model:
-    opt_loss = copy.deepcopy(loss_ema)
-    for save_dir in save_dir_all: 
-        torch.save(ddpm.state_dict(), save_dir + "model_initial.pth")
+import os
+
+# Check if model already exists
+model_exists = os.path.exists('./data/t500/model_initial.pth')
+
+if model_exists:
+    print("Found existing model_initial.pth - Loading and skipping training!")
+    ddpm.load_state_dict(torch.load('./data/t500/model_initial.pth'))
+else:
+    print("No existing model found - Training from scratch...")
+    optim = torch.optim.Adam(ddpm.parameters(), lr=lrate)
+    opt_loss = 1
+    pbar = tqdm(range(n_epoch),leave = True)
+    for ep in pbar:
+        ddpm.train()
+
+                # linear lrate decay
+        optim.param_groups[0]['lr'] = lrate*(1-ep/n_epoch)
+
+        loss_ema = None
+        for x, c in dataloader:
+            optim.zero_grad()
+            x = x.to(device)
+            c = c.to(device)
+            loss = ddpm(x, c, 1, 1, guide_w=0)
+            loss = loss.sum()
+            loss.backward()
+            if loss_ema is None:
+                loss_ema = loss.item()
+            else:
+                loss_ema = 0.95 * loss_ema + 0.05 * loss.item()
+                pbar.set_description(f"loss: {loss_ema:.4f}")
+                optim.step()
+
+    if save_model:
+        opt_loss = copy.deepcopy(loss_ema)
+        for save_dir in save_dir_all:
+            torch.save(ddpm.state_dict(), save_dir + "model_initial.pth")
+        print("Model saved!")
